+++ android/app/src/main/java/com/example/kurush_frontend/matching/main/MatchingCountryFragment.kt
@@ -1,14 +1,13 @@
-package com.example.kurush_frontend.matching
+package com.example.kurush_frontend.matching.main
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.kurush_frontend.R
 import com.example.kurush_frontend.databinding.FragmentMatchingMainBinding
 
-class MatchingMainFragment : Fragment() {
+class MatchingCountryFragment : Fragment() {
     lateinit var binding: FragmentMatchingMainBinding
 
     override fun onCreateView(
@@ -17,10 +16,7 @@ class MatchingMainFragment : Fragment() {
     ): View? {
         binding = FragmentMatchingMainBinding.inflate(inflater, container, false)
 
-
-
         return binding.root
     }
-
 
 }
