+++ android/app/src/main/java/com/example/kurush_frontend/matching/main/MatchingFreeFragment.kt
@@ -1,22 +1,22 @@
-package com.example.kurush_frontend.matching
+package com.example.kurush_frontend.matching.main
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.kurush_frontend.R
 import com.example.kurush_frontend.databinding.FragmentMatchingMainBinding
 
-class MatchingStudyFragment : Fragment() {
+class MatchingFreeFragment : Fragment() {
     lateinit var binding: FragmentMatchingMainBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentMatchingMainBinding.inflate(inflater, container,false)
+        binding = FragmentMatchingMainBinding.inflate(inflater, container, false)
 
         return binding.root
     }
+
 }
