+++ android/app/src/main/java/com/example/kurush_frontend/matching/main/MatchingHabitFragment.kt
@@ -1,11 +1,10 @@
-package com.example.kurush_frontend.matching
+package com.example.kurush_frontend.matching.main
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.kurush_frontend.R
 import com.example.kurush_frontend.databinding.FragmentMatchingMainBinding
 
 class MatchingHabitFragment : Fragment() {
