+++ android/app/src/main/java/com/example/kurush_frontend/matching/main/MatchingMainFragment.kt
@@ -1,14 +1,13 @@
-package com.example.kurush_frontend.matching
+package com.example.kurush_frontend.matching.main
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.kurush_frontend.R
 import com.example.kurush_frontend.databinding.FragmentMatchingMainBinding
 
-class MatchingFreeFragment : Fragment() {
+class MatchingMainFragment : Fragment() {
     lateinit var binding: FragmentMatchingMainBinding
 
     override fun onCreateView(
@@ -17,7 +16,10 @@ class MatchingFreeFragment : Fragment() {
     ): View? {
         binding = FragmentMatchingMainBinding.inflate(inflater, container, false)
 
+
+
         return binding.root
     }
+
 
 }
