+++ android/app/src/main/res/layout/fragment_matching_country.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".matching.MatchingCountryFragment">
+    tools:context=".matching.main.MatchingCountryFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
