+++ android/app/src/main/res/layout/fragment_matching_main.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".matching.MatchingMainFragment"
+    tools:context=".matching.main.MatchingMainFragment"
     android:background="#F5F5F5">
 
     <EditText
